WolfSSLSecureEndPoint: Modified the interface to match SecureEndPoint's interface

+++ include/TLSAbstractionLayer/WolfSSLSecureEndPoint.hpp
@@ -7,8 +7,8 @@
 
 /* WolfSSL */
 #include <wolfssl/options.h>
-#include <wolfssl/internal.h>
 #include <wolfssl/ssl.h>
+#include <wolfssl/openssl/ssl.h>
 
 /* Utilities */
 #include <sys/socket.h>
@@ -53,10 +53,10 @@ namespace TLSAbstractionLayer {
     int setupTLS();
     int setupIO(IO);
     int doHandshake();
-    int send(const char *, size_t);
-    int receive(char *, size_t);
-    int writeToBuffer(const char *,size_t size, char **);
-    int readFromBuffer(const char *,size_t size,char **);
+    int send(const char *, int);
+    int receive(char *, int);
+    int writeToBuffer(const char *,int size, char **);
+    int readFromBuffer(const char *,int size,char **);
   private:
     int setupProtocolAndVersion();
     int setupPeerVerification();
